aruha-600: changed layering
